fix "function return type inconsistent"
MISRA C required function return type should be consistented.

Acked-by: Eddie Dong <eddie.dong@intel.com>

+++ hypervisor/include/common/schedule.h
@@ -23,7 +23,7 @@ void get_schedule_lock(uint16_t pcpu_id);
 void release_schedule_lock(uint16_t pcpu_id);
 
 void set_pcpu_used(uint16_t pcpu_id);
-int allocate_pcpu(void);
+uint16_t allocate_pcpu(void);
 void free_pcpu(uint16_t pcpu_id);
 
 void add_vcpu_to_runqueue(struct vcpu *vcpu);
